requirements.txt
hikari[speedups,server]==2.0.0.dev115
python-dotenv~=0.21.0